+++ CHANGELOG.md
@@ -7,10 +7,11 @@
 - All commands require `context.Context` as a first argument, e.g. `rdb.Ping(ctx)`. If you are not
   using `context.Context` yet, the simplest option is to define global package variable
   `var ctx = context.TODO()` and use it when `ctx` is required.
-- `Cluster.ForEachNode` is renamed to `ForEachShard` for consistency with `Ring`.
 
 - Added `redisext.OpenTemetryHook` that adds
   [Redis OpenTelemetry instrumentation](https://redis.uptrace.dev/tracing/).
+
+- Redis slow log support.
 
 - Ring uses Rendezvous Hashing by default which provides better distribution. You need to move
   existing keys to a new location or keys will be inaccessible / lost. To use old hashing scheme:
@@ -24,6 +25,8 @@ ring := redis.NewRing(&redis.RingOptions{
     },
 })
 ```
+
+- `Cluster.ForEachNode` is renamed to `ForEachShard` for consistency with `Ring`.
 
 ## v7.3